requirements.txt
# AI Image Generation Evaluation Tool Dependencies

# Core dependencies
fal-client>=0.7.0
pyyaml>=6.0
pandas>=1.5.0

# Optional dependencies for enhanced functionality
requests>=2.28.0
pillow>=9.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
